requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
filelock==3.0.12
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
Markdown==3.3.3
mock==4.0.2
more-itertools==8.4.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
Pillow==8.0.1
pipenv==2020.6.2
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==5.4.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.25.9
virtualenv==20.0.27
virtualenv-clone==0.5.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
